be-py/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
google-genai
ollama
